[CALCITE-1050] Array support for Avatica
As best as possible, works around the limitations of JDBC's
Array class to handle arbitrarily nested Arrays. Nested-array
support differs from DB to DB, so functionality is primarily
driven from structure, rather than metadata.

Also, fixes the implementation of Array.getArray(long, int).The
arguments are given in terms of one-offset and number of elements,
whereas the list.subList(int, int) call is expecting zero-offset and
exclusive end-offset.

Reduce the "API" changes to avoid downstream changes in Avatica. This
"restores" some weirdness in the implementations, but simplifies the
upgrade process (probably worth it).

Closes apache/calcite-avatica#2

Change-Id: I88e3fccb784c6673523ebc23896e2e9a197c4150